debugged behaviors, added logger, added window border

+++ inc/cmath.h
@@ -12,7 +12,7 @@ public:
 
     static void clamp(int& value, int min, int max);
 
-    static float sqrDistance(int x1, int y1, int x2, int y2);
+    static int sqrDistance(int x1, int y1, int x2, int y2);
 
     static float sqrt(float value);
